Update the abqpy structure
